+++ apps/web/src/app/(main)/(docs)/components/header/variants/default.tsx
@@ -2,8 +2,8 @@ import { Button, Header, type HeaderProps, Repeater } from "@jamsr-ui/react";
 
 export const HeaderDefault = (props: HeaderProps) => {
   return (
-    <div>
-      <Header className="flex justify-between px-2" {...props}>
+    <>
+      <Header className="flex justify-between px-4" {...props}>
         {/* eslint-disable-next-line @next/next/no-img-element */}
         <img
           className="h-10"
@@ -12,8 +12,8 @@ export const HeaderDefault = (props: HeaderProps) => {
         />
         <Button>Go to dashboard</Button>
       </Header>
-      <div className="py-4">
-        <Repeater count={212}>
+      <div className="flex flex-col items-center justify-center gap-8 p-4">
+        <Repeater count={12}>
           <p>
             Lorem ipsum dolor sit amet consectetur adipisicing elit. Fugit
             nobis, repudiandae ullam animi illum praesentium quidem eius amet
@@ -22,6 +22,6 @@ export const HeaderDefault = (props: HeaderProps) => {
           </p>
         </Repeater>
       </div>
-    </div>
+    </>
   );
 };