+++ apps/web/src/app/(main)/(docs)/components/header/variants/static.tsx
@@ -1,0 +1,5 @@
+import { HeaderDefault } from "./default";
+
+export const HeaderStatic = () => {
+  return <HeaderDefault position="static" />;
+};