+++ apps/web/src/app/(main)/(docs)/components/header/variants/with-border.tsx
@@ -1,0 +1,5 @@
+import { HeaderDefault } from "./default";
+
+export const HeaderWithBorder = () => {
+  return <HeaderDefault isBordered />;
+};