+++ apps/web/src/app/(main)/layout.tsx
@@ -1,0 +1,17 @@
+import { AppHeader } from "@/layouts/app-header";
+import React from "react";
+
+type Props = {
+  children: React.ReactNode;
+};
+const Layout = (props: Props) => {
+  const { children } = props;
+  return (
+    <>
+      <AppHeader />
+      {children}
+    </>
+  );
+};
+
+export default Layout;