+++ apps/web/src/app/examples/header/default/page.tsx
@@ -1,0 +1,7 @@
+import { HeaderDefault } from "@/app/(main)/(docs)/components/header/variants/default";
+
+const Page = () => {
+  return <HeaderDefault />;
+};
+
+export default Page;