+++ apps/web/src/app/examples/header/hide-on-scroll/page.tsx
@@ -1,0 +1,7 @@
+import { HeaderHideOnScroll } from "@/app/(main)/(docs)/components/header/variants/hide-on-scroll";
+
+const Page = () => {
+  return <HeaderHideOnScroll />;
+};
+
+export default Page;