+++ apps/web/src/app/examples/header/static/page.tsx
@@ -1,0 +1,7 @@
+import { HeaderStatic } from "@/app/(main)/(docs)/components/header/variants/static";
+
+const Page = () => {
+  return <HeaderStatic />;
+};
+
+export default Page;