+++ apps/web/src/app/examples/header/with-border/page.tsx
@@ -1,0 +1,7 @@
+import { HeaderWithBorder } from "@/app/(main)/(docs)/components/header/variants/with-border";
+
+const Page = () => {
+  return <HeaderWithBorder />;
+};
+
+export default Page;